restructure examples
