+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members  = ["crates/*", "examples/*"]
+members  = ["crates/*", "examples"]
 resolver = "3"
 
 [workspace.package]
