+++ crates/aeronet_steam/examples/steam_client.rs
@@ -95,7 +95,7 @@ fn global_ui(
     mut target_peer: Local<String>,
     mut session_id: Local<usize>,
 ) {
-    const DEFAULT_TARGET: &str = "127.0.0.1:25567";
+    const DEFAULT_TARGET: &str = "127.0.0.1:25572";
 
     egui::Window::new("Connect").show(egui.ctx_mut(), |ui| {
         let enter_pressed = ui.input(|i| i.key_pressed(egui::Key::Enter));
