+++ crates/aeronet_webtransport/examples/webtransport_server.rs
@@ -56,7 +56,7 @@ fn open_server(mut commands: Commands) {
 
 fn server_config(identity: wtransport::Identity) -> ServerConfig {
     wtransport::ServerConfig::builder()
-        .with_bind_default(25565)
+        .with_bind_default(25571)
         .with_identity(identity)
         .keep_alive_interval(Some(Duration::from_secs(1)))
         .max_idle_timeout(Some(Duration::from_secs(5)))
