--- examples/echo_client/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-description = "Demo client which connects to an echo server"
-name        = "echo_client"
-publish     = false
-
-authors.workspace      = true
-categories.workspace   = true
-edition.workspace      = true
-keywords.workspace     = true
-license.workspace      = true
-repository.workspace   = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-aeronet           = { workspace = true }
-aeronet_websocket = { workspace = true, features = ["client"] }
-bevy              = { workspace = true }
-bevy_egui         = { workspace = true }
