--- examples/echo_server/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-description = "Demo server which echoes back what clients send it"
-name        = "echo_server"
-publish     = false
-
-authors.workspace      = true
-categories.workspace   = true
-edition.workspace      = true
-keywords.workspace     = true
-license.workspace      = true
-repository.workspace   = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-aeronet           = { workspace = true }
-aeronet_websocket = { workspace = true, features = ["server"] }
-bevy              = { workspace = true }
