--- examples/move_box/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-description = "Demo app with moving boxes synced using bevy_replicon"
-name        = "move_box"
-publish     = false
-
-authors.workspace      = true
-categories.workspace   = true
-edition.workspace      = true
-keywords.workspace     = true
-license.workspace      = true
-repository.workspace   = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-bevy          = { workspace = true, features = ["wayland"] }
-bevy_replicon = { workspace = true }
-serde         = { workspace = true }
