--- examples/move_box_client/README.md
@@ -1,1 +1,0 @@
-See [`move_box`](../move_box).
