--- examples/move_box_server/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-description = "move_box demo server"
-name        = "move_box_server"
-publish     = false
-
-authors.workspace      = true
-categories.workspace   = true
-edition.workspace      = true
-keywords.workspace     = true
-license.workspace      = true
-repository.workspace   = true
-rust-version.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-move_box = { path = "../move_box" }
-
-aeronet              = { workspace = true }
-aeronet_replicon     = { workspace = true, features = ["server"] }
-aeronet_websocket    = { workspace = true, features = ["server"] }
-aeronet_webtransport = { workspace = true, features = ["server"] }
-bevy                 = { workspace = true }
-bevy_replicon        = { workspace = true }
-clap                 = { workspace = true }
