+++ examples/src/lib.rs
@@ -1,0 +1,3 @@
+//! Shared code for examples.
+
+pub mod move_box;
